The comment selection has been expanded. The tokenizer is now newline-aware.
The incoming string is first transformed as follows:

  ~~ --> ~~E
  \n --> ~~M

The string is then tokenized. In the case of a substitution or the (js)
command, the string is decoded prior to the evaluation and the result is
reencoded again afterwards. The string is also decoded prior to being
sent / logged to the console.

This solution may seem like an overkill, but none of the regexes used
throughout the tokenizer and the execution pipeline codebase are
multiline and it's much easier to transform all incoming strings into
single-line strings than to rewrite all of the regexes.